// frontend/src/ipc/generated/CalendarDayProjection.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CalendarDayProjection = { date: string, is_today: boolean, is_selected: boolean, task_count: number, scheduled_minutes: number, category_icon_keys: Array<string>, extra_category_count: number, morning_load_ratio: number, afternoon_load_ratio: number, evening_load_ratio: number, has_missed: boolean, };